♻️ queries/hooks/mogaco를 queries/hooks/post로 이동
