+++ app/frontend/src/queries/hooks/index.ts
@@ -1,2 +1,1 @@
 export * from './member';
-export * from './mogaco';
